divided endpoints from cart from orders and ordermanagement routes, refactored API endpoint names in backend and frontend both
